# docs/swagger/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "SlackCheers API for workspace setup, people management, channel settings, and celebrations.",
        "title": "SlackCheers API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/auth/slack/callback": {
            "get": {
                "description": "Exchanges OAuth code, stores workspace install metadata, and returns connected workspace details.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Slack OAuth callback",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Slack OAuth code",
                        "name": "code",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Opaque CSRF state",
                        "name": "state",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Slack OAuth error",
                        "name": "error",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_http_handlers.SlackConnectResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_http_handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_http_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/slack/install": {
            "get": {
                "description": "Redirects to Slack OAuth consent page. Use mode=json to return URL without redirect.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Start Slack install",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Opaque CSRF state",
                        "name": "state",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Set to json to return install URL",
                        "name": "mode",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_http_handlers.SlackInstallURLResponse"
                        }
                    },
                    "307": {
                        "description": "Temporary Redirect",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_http_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/slack/events": {
            "post": {
                "description": "Verifies Slack signatures, handles URL verification, and processes DM replies to save birthdays/hire dates.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "slack"
                ],
                "summary": "Slack events webhook",
                "parameters": [
                    {
                        "description": "Slack event payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_http_handlers.SlackEventEnvelope"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_http_handlers.SlackEventAckResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_http_handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_http_handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_http_handlers.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "internal_http_handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "internal_http_handlers.SlackConnectResponse": {
            "type": "object",
            "properties": {
                "installation": {
                    "$ref": "#/definitions/internal_http_handlers.SlackOAuthInstallation"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "internal_http_handlers.SlackEventAckResponse": {
            "type": "object",
            "properties": {
                "challenge": {
                    "type": "string"
                },
                "ok": {
                    "type": "boolean"
                }
            }
        },
        "internal_http_handlers.SlackEventEnvelope": {
            "type": "object",
            "properties": {
                "challenge": {
                    "type": "string"
                },
                "event": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "team_id": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "internal_http_handlers.SlackInstallURLResponse": {
            "type": "object",
            "properties": {
                "install_url": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                }
            }
        },
        "internal_http_handlers.SlackOAuthInstallation": {
            "type": "object",
            "properties": {
                "bot_user_id": {
                    "type": "string"
                },
                "scope": {
                    "type": "string"
                },
                "team_id": {
                    "type": "string"
                },
                "team_name": {
                    "type": "string"
                },
                "workspace_id": {
                    "type": "string"
                }
            }
        },
...
    }
}